+++ websocket/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/websocket/route/WebSocketMiddleware.kt
@@ -1,10 +1,13 @@
-package lt.saltyjuice.dragas.chatty.websocket
+package lt.saltyjuice.dragas.chatty.v3.websocket.route
 
 import lt.saltyjuice.dragas.chatty.v3.core.middleware.AfterMiddleware
 import lt.saltyjuice.dragas.chatty.v3.core.middleware.BeforeMiddleware
-import lt.saltyjuice.dragas.chatty.websocket.message.Request
-import lt.saltyjuice.dragas.chatty.websocket.message.Response
+import lt.saltyjuice.dragas.chatty.v3.websocket.message.Request
+import lt.saltyjuice.dragas.chatty.v3.websocket.message.Response
 
+/**
+ * WebSocket wrapper for usual middleware.
+ */
 open class WebSocketMiddleware : BeforeMiddleware<Request>, AfterMiddleware<Response>
 {
     override fun before(request: Request): Boolean